LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/eligibility/inflight-fetch-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Invalid URL

Harness Error (TIMEOUT), message = null

TIMEOUT Eligibility (in-flight fetch): Header received before BFCache and body received when in BFCache Test timed out
NOTRUN Eligibility (in-flight fetch): Header received before BFCache and body received after BFCache
NOTRUN Eligibility (in-flight fetch): Header and body received when in BFCache
NOTRUN Eligibility (in-flight fetch): Header received when in BFCache and body received after BFCache
NOTRUN Eligibility (in-flight fetch): Header and body received after BFCache
